Draft: move more functions to draftgeoutils.general
